Back out "Use ConcreteStateTeller in RCTParagraphComponentView"
Summary:
RCTParagraphComponentView does not need to have the state teller because it does not change the state.
Also, we are thinking about removing State Teller, so to simplify a fix coming in the next diff, we are backout the introducing a state telled in RCTParagraphComponentView.

Changelog: [Internal] Fabric-specific internal change.

Reviewed By: sammy-SC

Differential Revision: D24095655

fbshipit-source-id: 90bd21cb8b01056c0db902a604b4088bf8c7750e